requirements.txt
# Core FastAPI stack
fastapi>=0.116.1
uvicorn>=0.35.0
gunicorn>=23.0.0
python-multipart>=0.0.20

# Database and API clients
supabase>=1.0.3
openai>=1.3.0
httpx>=0.23.3

# Configuration
python-dotenv>=1.0.0
pydantic>=2.11.7

# ML libraries (current versions)
sentence-transformers>=5.1.0
scikit-learn>=1.7.1
torch>=2.8.0
numpy>=2.3.2

# Document processing
python-docx>=1.1.0
openpyxl>=3.1.2
python-dateutil>=2.9.0

# Essential dependencies
requests>=2.31.0
typing-extensions>=4.14.1
